change id data type as int

+++ backend/eventsbooking/src/main/java/com/application/eventsbooking/models/BusinessEntity.java
@@ -7,7 +7,7 @@ public abstract class BusinessEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    protected Long id;
+    protected int id;
 
     @Column(nullable = false)
     protected String name;
